Added Chronos Python Script
Developed Chronos Python script for extracting raw accelerometer data.
Compiled with VS 15. Can be run by opening command prompt and typing in
the just the location of the python source file. It will then create an
ACCEL.csv file in the folder where the source file is contianed. This
ACCEL.csv file contains the acceleration data captured from the Chronos
watch. Only the data where a single coordinate (x,y or z) is greater
than 0 will be recorded. If all are zero nothing will be saved. Each
record is time stamped in seconds for now.